engram: q_weight/k_weight are bf16, and f32() now refuses anything else
The container keeps the checkpoint's bf16 for engram_q and engram_k, but the
loader read them through f32(), which reinterpreted the bytes as floats --
and twice as many as the tensor holds, running into the next tensor. The
gate's q*k weight was garbage, every gate sat near 0.5, and engram wrote ~60x
the reference's energy into the residual at layers 1 and 14 (measured on a
16-token prompt against the reference with engram: +1.33e5 vs +2.1e3). The
result decoded fluently and ignored the prompt: the planted-fact needle
failed 0/2 at 4K and 16K with engram on and passed 2/2 with it off.

Read both as bf16 and convert. f32() now refuses a non-F32 tensor exactly as
bf16() refuses a non-BF16 one -- this was the fourth V4.1 tensor to fall
through a V4-era loader rule, and the only one of the f32() reads that was
not f32 (checked against aff-info for every name).

After this and e03a05b (no q RMS on V4.1): needle 2/2 at 4K and 16K with
engram ON.

+++ src/engine/model.h
@@ -180,8 +180,8 @@ struct LayerWeights {
   // Engram, on `engram_layers` only: q/k are [hc_mult, n_embd] mixes and wkv projects the looked-up
   // n-gram rows. The TABLE itself is not in the container.
   DenseW engram_wkv ;
-  const float*    engram_q = nullptr;
-  const float*    engram_k = nullptr;
+  const uint16_t* engram_q = nullptr;   // bf16 in the container, as in the checkpoint
+  const uint16_t* engram_k = nullptr;
 
   // Shared expert — dense, runs on every token alongside the routed six.
   DenseW shexp_gate ;
